+++ examples/basic-sticky.rs
@@ -3,18 +3,21 @@ use std::thread;
 use fragile::Sticky;
 
 fn main() {
-    fragile::stack_token!(tok);
-
-    // creating and using a fragile object in the same thread works
+    // creating and using a sticky object in the same thread works
     let val = Sticky::new(true);
-    println!("debug print in same thread: {:?}", &val);
-    println!("try_get in same thread: {:?}", val.try_get(tok));
+    println!("debug print in same thread: {:?}", val);
+    println!(
+        "try_with in same thread: {:?}",
+        val.try_with(|value| *value)
+    );
 
-    // once send to another thread it stops working
+    // once sent to another thread it stops working
     thread::spawn(move || {
-        fragile::stack_token!(tok);
-        println!("debug print in other thread: {:?}", &val);
-        println!("try_get in other thread: {:?}", val.try_get(tok));
+        println!("debug print in other thread: {:?}", val);
+        println!(
+            "try_with in other thread: {:?}",
+            val.try_with(|value| *value)
+        );
     })
     .join()
     .unwrap();